requirements.txt
python-dotenv>=1.0.0
requests>=2.25.0
genanki>=0.13.0
